+++ crates/sml-core/Cargo.toml
@@ -9,3 +9,4 @@ edition = "2018"
 [dependencies]
 sml-util = {path = "../sml-util"}
 sml-frontend = {path = "../sml-frontend"}
+typed-arena = "2.0"
