+++ crates/sml-core/src/lib.rs
@@ -1,14 +1,18 @@
 use sml_frontend::ast::Const;
+use sml_util::arena::Arena;
 use sml_util::interner::Symbol;
 use sml_util::span::Span;
 use std::collections::HashMap;
 use std::fmt;
 
+pub mod arenas;
 pub mod builtin;
 pub mod check;
 pub mod elaborate;
 pub mod types;
 use types::{Constructor, Scheme, Tycon, Type, TypeVar};
+
+pub use arenas::CoreArena;
 
 #[derive(Copy, Clone, Debug, Default, PartialEq, PartialOrd, Eq, Hash)]
 pub struct TypeId(pub u32);
@@ -16,47 +20,47 @@ pub struct TypeId(pub u32);
 #[derive(Copy, Clone, Debug, Default, PartialEq, PartialOrd, Eq, Hash)]
 pub struct ExprId(pub u32);
 
-#[derive(Clone)]
-pub enum ExprKind {
-    App(Box<Expr>, Box<Expr>),
-    Case(Box<Expr>, Vec<Rule>),
-    Con(Constructor, Vec<Type>),
+// #[derive(Clone)]
+pub enum ExprKind<'ar> {
+    App(Expr<'ar>, Expr<'ar>),
+    Case(Expr<'ar>, Vec<Rule<'ar>>),
+    Con(Constructor, Vec<&'ar Type<'ar>>),
     Const(Const),
-    Handle(Box<Expr>, Vec<Rule>),
-    Lambda(Box<Lambda>),
-    Let(Vec<Decl>, Box<Expr>),
-    List(Vec<Expr>),
+    Handle(Expr<'ar>, Vec<Rule<'ar>>),
+    Lambda(Lambda<'ar>),
+    Let(Vec<Decl<'ar>>, Expr<'ar>),
+    List(Vec<Expr<'ar>>),
     Primitive(Symbol),
-    Raise(Box<Expr>),
-    Record(Vec<Row<Expr>>),
-    Seq(Vec<Expr>),
+    Raise(Expr<'ar>),
+    Record(Vec<Row<Expr<'ar>>>),
+    Seq(Vec<Expr<'ar>>),
     Var(Symbol),
 }
 
-#[derive(Clone)]
-pub struct Expr {
-    pub expr: ExprKind,
-    pub ty: Type,
+#[derive(Copy, Clone)]
+pub struct Expr<'ar> {
+    pub expr: &'ar ExprKind<'ar>,
+    pub ty: &'ar Type<'ar>,
     pub span: Span,
 }
 
 #[derive(Clone)]
-pub struct Lambda {
+pub struct Lambda<'ar> {
     pub arg: Symbol,
-    pub ty: Type,
-    pub body: Expr,
+    pub ty: &'ar Type<'ar>,
+    pub body: Expr<'ar>,
 }
 
 #[derive(Clone)]
-pub enum PatKind {
+pub enum PatKind<'ar> {
     /// Constructor application
-    App(Constructor, Option<Box<Pat>>),
+    App(Constructor, Option<Pat<'ar>>),
     /// Constant
     Const(Const),
     /// Literal list
-    List(Vec<Pat>),
+    List(Vec<Pat<'ar>>),
     /// Record
-    Record(Vec<Row<Pat>>),
+    Record(Vec<Row<Pat<'ar>>>),
     /// Variable binding
     Var(Symbol),
     /// Wildcard
@@ -64,23 +68,23 @@ pub enum PatKind {
 }
 
 #[derive(Clone)]
-pub struct Pat {
-    pub pat: PatKind,
-    pub ty: Type,
+pub struct Pat<'ar> {
+    pub pat: &'ar PatKind<'ar>,
+    pub ty: &'ar Type<'ar>,
     pub span: Span,
 }
 
 #[derive(Clone, Debug)]
-pub struct Rule {
-    pub pat: Pat,
-    pub expr: Expr,
+pub struct Rule<'ar> {
+    pub pat: Pat<'ar>,
+    pub expr: Expr<'ar>,
 }
 
 #[derive(Clone, Debug)]
-pub struct Datatype {
-    tycon: Tycon,
-    tyvars: Vec<usize>,
-    constructors: Vec<(Constructor, Option<Type>)>,
+pub struct Datatype<'ar> {
+    pub tycon: Tycon,
+    pub tyvars: Vec<usize>,
+    pub constructors: Vec<(Constructor, Option<&'ar Type<'ar>>)>,
 }
 
 // #[derive(Clone, Debug)]
@@ -91,22 +95,22 @@ pub struct Datatype {
 // }
 
 #[derive(Clone, Debug)]
-pub enum Decl {
-    Datatype(Datatype),
-    Fun(Vec<usize>, Vec<Lambda>),
-    Val(Rule),
-    Exn(Constructor, Option<Type>),
+pub enum Decl<'ar> {
+    Datatype(Datatype<'ar>),
+    Fun(Vec<usize>, Vec<Lambda<'ar>>),
+    Val(Rule<'ar>),
+    Exn(Constructor, Option<&'ar Type<'ar>>),
 }
 
-#[derive(Clone, Debug, PartialEq, PartialOrd)]
+#[derive(Copy, Clone, Debug, PartialEq, PartialOrd)]
 pub struct Row<T> {
     pub label: Symbol,
     pub data: T,
     pub span: Span,
 }
 
-impl Expr {
-    pub fn new(expr: ExprKind, ty: Type, span: Span) -> Expr {
+impl<'ar> Expr<'ar> {
+    pub fn new(expr: &'ar ExprKind<'ar>, ty: &'ar Type<'ar>, span: Span) -> Expr<'ar> {
         Expr { expr, ty, span }
     }
 
@@ -125,18 +129,18 @@ impl Expr {
     }
 }
 
-impl Pat {
-    pub fn new(pat: PatKind, ty: Type, span: Span) -> Pat {
+impl<'ar> Pat<'ar> {
+    pub fn new(pat: &'ar PatKind<'ar>, ty: &'ar Type<'ar>, span: Span) -> Pat<'ar> {
         Pat { pat, ty, span }
     }
 }
 
 impl<T> Row<T> {
-    pub fn fmap<S, F: FnOnce(T) -> S>(self, f: F) -> Row<S> {
+    pub fn fmap<S, F: FnOnce(&T) -> S>(&self, f: F) -> Row<S> {
         Row {
             label: self.label,
             span: self.span,
-            data: f(self.data),
+            data: f(&self.data),
         }
     }
 }
@@ -150,7 +154,7 @@ impl<T, E> Row<Result<T, E>> {
     }
 }
 
-impl fmt::Debug for ExprKind {
+impl<'ar> fmt::Debug for ExprKind<'ar> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         use ExprKind::*;
         match self {
@@ -194,7 +198,7 @@ impl fmt::Debug for ExprKind {
     }
 }
 
-impl fmt::Debug for PatKind {
+impl<'ar> fmt::Debug for PatKind<'ar> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         use PatKind::*;
         match self {
@@ -216,19 +220,19 @@ impl fmt::Debug for PatKind {
     }
 }
 
-impl fmt::Debug for Expr {
+impl<'ar> fmt::Debug for Expr<'ar> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "[{:?} : {:?}]", self.expr, self.ty)
     }
 }
 
-impl fmt::Debug for Pat {
+impl<'ar> fmt::Debug for Pat<'ar> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{:?}", self.pat)
     }
 }
 
-impl fmt::Debug for Lambda {
+impl<'ar> fmt::Debug for Lambda<'ar> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "fn {:?} => {:?}", self.arg, self.body)
     }
