+++ crates/sml-util/src/arena.rs
@@ -1,7 +1,7 @@
 use std::cell::RefCell;
 use std::mem;
 
-const PAGE_SIZE: usize = 4096;
+const PAGE_SIZE: usize = 8192*8;
 
 pub struct Arena<T> {
     chunks: RefCell<Chunks<T>>,
@@ -57,7 +57,7 @@ impl<T> Arena<T> {
 
 impl<T> Chunks<T> {
     fn reserve(&mut self) {
-        let cap = self.current.capacity() * 2;
+        let cap = self.current.capacity();
         let chunk = mem::replace(&mut self.current, Vec::with_capacity(cap));
         self.rest.push(chunk);
     }
